backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pydantic==2.10.1
pydantic-settings==2.6.1
pydantic[email]==2.10.1
python-jose[cryptography]==3.3.0
httpx==0.28.1
google-generativeai==0.8.3
google-cloud-speech==2.28.0
google-cloud-storage==2.19.0
google-auth==2.36.0
playwright==1.49.0
upstash-redis==1.3.0
pinecone==5.4.2
sentry-sdk[fastapi]==2.19.2
svix==1.24.0
aiofiles==24.1.0
python-multipart==0.0.12
